Support fractional seconds for delay
Allow power cycle delay to be fractional seconds, e.g. not only 2 sec,
but also like 0.3 sec or 1.5 sec.

+++ uhubctl.c
@@ -200,7 +200,7 @@ static char opt_vendor[16]   = "";
 static char opt_location[32] = "";     /* Hub location a-b.c.d */
 static int opt_ports  = ALL_HUB_PORTS; /* Bitmask of ports to operate on */
 static int opt_action = POWER_KEEP;
-static int opt_delay  = 2;
+static double opt_delay = 2;
 static int opt_repeat = 1;
 static int opt_wait   = 20; /* wait before repeating in ms */
 static int opt_exact  = 0;  /* exact location match - disable USB3 duality handling */
@@ -234,7 +234,7 @@ static int print_usage()
         "--ports,    -p - ports to operate on    [all hub ports].\n"
         "--loc,      -l - limit hub by location  [all smart hubs].\n"
         "--vendor,   -n - limit hub by vendor id [%s] (partial ok).\n"
-        "--delay,    -d - delay for cycle action [%d sec].\n"
+        "--delay,    -d - delay for cycle action [%g sec].\n"
         "--repeat,   -r - repeat power off count [%d] (some devices need it to turn off).\n"
         "--exact,    -e - exact location (no USB3 duality handling).\n"
         "--reset,    -R - reset hub after each power-on action, causing all devices to reassociate.\n"
@@ -796,7 +796,7 @@ int main(int argc, char *argv[])
             }
             break;
         case 'd':
-            opt_delay = atoi(optarg);
+            opt_delay = atof(optarg);
             break;
         case 'r':
             opt_repeat = atoi(optarg);
@@ -962,7 +962,7 @@ int main(int argc, char *argv[])
             libusb_close(devh);
         }
         if (k == 0 && opt_action == POWER_CYCLE)
-            sleep_ms(opt_delay * 1000);
+            sleep_ms((int)(opt_delay * 1000));
     }
     rc = 0;
 cleanup:
